Simple pallet config
